# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Sheba e-Government API
  version: "1.0.0"
  description: >
    Yemen's national e-Government backend — OIDC/OAuth 2.1 identity provider (OpenIddict) and
    government service gateway. **Generated from the current implementation** (`src/`). Every REST
    response (except `/connect/*` and `/.well-known/*`) is wrapped in a JSend envelope
    (`{status, data, message?, code?}`). OIDC endpoints speak OAuth error JSON, not JSend.
    See docs/frontend-api.md for the human-readable reference and known discrepancies.
  contact:
    name: Sheba Platform Team
    email: support@sheba.gov
servers:
  - url: https://localhost:7001
    description: Local dev (HTTP allowed in dev)
...
security:
  - bearerAuth: []
paths:
...
  /api/ministry:
    get:
      tags: [Ministry]
      summary: "List ministries"
      parameters: [ { name: includeInactive, in: query, schema: { type: boolean, default: false } } ]
      responses: { '200': { description: Ministries, content: { application/json: { schema: { $ref: '#/components/schemas/JSend_MinistryList' } } } } }
    post:
      tags: [Ministry]
      summary: "Create ministry"
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateMinistryCommand' } } } }
      responses: { '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/JSendSuccess' } } } } }
  /api/ministry/{id}:
    get:
      tags: [Ministry]
      summary: "Ministry detail"
      parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses:
        '200': { description: Detail, content: { application/json: { schema: { $ref: '#/components/schemas/JSend_MinistryDetail' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Ministry]
      summary: "Update ministry"
      parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateMinistryCommand' } } } }
      responses: { '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/JSendSuccess' } } } } }
  /api/ministry/{id}/auth-config:
    post:
      tags: [Ministry]
      summary: "Add auth configuration (encrypted credentials)"
      parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/SetMinistryAuthConfigCommand' } } } }
      responses: { '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/JSendSuccess' } } } } }
  /api/ministry/{id}/test-connection:
    post:
      tags: [Ministry]
      summary: "Test connectivity to a ministry API"
      parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { authConfigId: { type: string, format: uuid } }, required: [authConfigId] } } } }
      responses: { '200': { description: Result, content: { application/json: { schema: { $ref: '#/components/schemas/JSendSuccess' } } } } }
  /api/ministry/{id}/endpoints:
    post:
      tags: [Ministry]
      summary: "Register a ministry endpoint"
      parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/RegisterMinistryEndpointCommand' } } } }
      responses: { '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/JSendSuccess' } } } } }
  /api/ministry/{id}/webhooks:
    post:
      tags: [Ministry]
      summary: "Register a ministry webhook"
      parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/RegisterMinistryWebhookCommand' } } } }
      responses: { '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/JSendSuccess' } } } } }

  # ─────────────────────────── Service Catalog ───────────────────────────
...
  /api/webhooks/ministry/{ministryId}:
    post:
      tags: [Webhooks]
      summary: "Inbound ministry callback (HMAC-verified) — not a frontend surface"
      security: []
      parameters:
        - { name: ministryId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: X-Sheba-Event, in: header, schema: { type: string } }
        - { name: X-Sheba-Signature, in: header, schema: { type: string } }
        - { name: X-Sheba-Timestamp, in: header, schema: { type: string } }
        - { name: X-Sheba-Delivery-Id, in: header, schema: { type: string } }
      requestBody: { content: { application/json: { schema: { type: object, additionalProperties: true } } } }
      responses:
        '200': { description: Accepted }
        '400': { description: Rejected (bad signature/timestamp/duplicate) }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token from /connect/token."
...
  responses:
...
    NotFound:
      description: "404 — resource not found"
      content: { application/json: { schema: { $ref: '#/components/schemas/JSendFail' } } }
...
  schemas:
...
    JSendSuccess:
      type: object
      properties:
        status: { type: string, const: success }
        data: {}
      required: [status]
    JSendFail:
      type: object
      properties:
        status: { type: string, const: fail }
        data:
          type: object
          additionalProperties: { type: string }
          description: "field → message dictionary"
      required: [status, data]
...
    MinistryAuthType: { type: string, enum: [Oidc, OAuth2, ApiKey, BearerToken, BasicAuth, Saml, Custom, None] }
    EndpointType: { type: string, enum: [DataQuery, ServiceAction, WebhookRegister, StatusCheck, DocumentFetch, Health] }
    ApiKeyPlacement: { type: string, enum: [Header, Query, Cookie] }
...
    CreateMinistryCommand:
      type: object
      properties:
        code: { type: string }
        nameAr: { type: string }
        nameEn: { type: string }
        parentMinistryId: { type: string, format: uuid, nullable: true }
        descriptionAr: { type: string, nullable: true }
        descriptionEn: { type: string, nullable: true }
        contactEmail: { type: string, nullable: true }
        contactPhone: { type: string, nullable: true }
      required: [code, nameAr, nameEn]
    UpdateMinistryCommand:
      type: object
      properties:
        nameAr: { type: string }
        nameEn: { type: string }
        descriptionAr: { type: string, nullable: true }
        descriptionEn: { type: string, nullable: true }
        logoUrl: { type: string, nullable: true }
        websiteUrl: { type: string, nullable: true }
        contactEmail: { type: string, nullable: true }
        contactPhone: { type: string, nullable: true }
        addressAr: { type: string, nullable: true }
        addressEn: { type: string, nullable: true }
        displayOrder: { type: integer, default: 0 }
        isActive: { type: boolean, default: true }
      required: [nameAr, nameEn]
    SetMinistryAuthConfigCommand:
      type: object
      properties:
        name: { type: string }
        authType: { $ref: '#/components/schemas/MinistryAuthType' }
        baseUrl: { type: string }
        isDefault: { type: boolean, default: false }
        healthCheckPath: { type: string, nullable: true }
        timeoutSeconds: { type: integer, default: 30 }
        retryCount: { type: integer, default: 3 }
        oidcTokenEndpoint: { type: string, nullable: true }
        oidcClientId: { type: string, nullable: true }
        oidcClientSecret: { type: string, nullable: true }
        oidcScope: { type: string, nullable: true }
        apiKeyHeaderName: { type: string, nullable: true }
        apiKeyValue: { type: string, nullable: true }
        apiKeyPlacementType: { $ref: '#/components/schemas/ApiKeyPlacement' }
        bearerToken: { type: string, nullable: true }
        basicUsername: { type: string, nullable: true }
        basicPassword: { type: string, nullable: true }
      required: [name, authType, baseUrl]
    RegisterMinistryEndpointCommand:
      type: object
      properties:
        code: { type: string }
        nameAr: { type: string }
        nameEn: { type: string }
        httpMethod: { type: string }
        pathTemplate: { type: string }
        type: { $ref: '#/components/schemas/EndpointType' }
        authConfigId: { type: string, format: uuid, nullable: true }
        descriptionAr: { type: string, nullable: true }
        descriptionEn: { type: string, nullable: true }
        timeoutSeconds: { type: integer, default: 30 }
        rateLimitPerMinute: { type: integer, nullable: true }
        requiresCitizenConsent: { type: boolean, default: false }
      required: [code, nameAr, nameEn, httpMethod, pathTemplate, type]
    RegisterMinistryWebhookCommand:
      type: object
      properties:
        eventType: { type: string }
        shebaWebhookPath: { type: string }
        signingSecret: { type: string }
        endpointId: { type: string, format: uuid, nullable: true }
      required: [eventType, shebaWebhookPath, signingSecret]
...
    MinistrySummaryDto:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        nameAr: { type: string }
        nameEn: { type: string }
        parentMinistryId: { type: string, format: uuid, nullable: true }
        depthLevel: { type: integer }
        isActive: { type: boolean }
        endpointCount: { type: integer }
        authConfigCount: { type: integer }
        createdAt: { type: string, format: date-time }
...
    JSend_MinistryList: { allOf: [ { $ref: '#/components/schemas/JSendSuccess' }, { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/MinistrySummaryDto' } } } } ] }
    JSend_MinistryDetail: { allOf: [ { $ref: '#/components/schemas/JSendSuccess' }, { type: object, properties: { data: { type: object, additionalProperties: true } } } ] }
...